requirements.txt
torch
torchvision
opencv-python
matplotlib
pyyaml
timm